requirements.txt
aiogram==3.x.x
python-dotenv
fpdf2
